# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sovren API
  description: |
    **Sovren API**: NOSTR-native decentralized creator monetization platform

    ## Overview
    The Sovren API provides comprehensive endpoints for content management, user profiles,
    and payment processing via Bitcoin Lightning Network. Built on the NOSTR protocol
    for true decentralization and censorship resistance.

    ## Base URLs
    - **Development**: `http://localhost:3001/api/v1`
    - **Production**: `https://api.sovren.app/v1`

    ## Authentication
    All authenticated endpoints require a JWT bearer token obtained via the authentication flow.
    The token must include the user's NOSTR public key for protocol compliance.

    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## Rate Limiting
    API requests are rate-limited per endpoint category:
    - Content operations: 100 requests/hour
    - User operations: 200 requests/hour
    - Payment operations: 50 requests/hour

    Rate limit information is returned in response headers:
    - `X-RateLimit-Limit`: Total requests allowed
    - `X-RateLimit-Remaining`: Requests remaining
    - `X-RateLimit-Reset`: Time when limit resets (Unix timestamp)

    ## Error Handling
    All errors follow a consistent format with appropriate HTTP status codes:
    - `400`: Bad Request - Validation error
    - `401`: Unauthorized - Authentication required
    - `403`: Forbidden - Insufficient permissions
    - `404`: Not Found - Resource doesn't exist
...

...

...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /payments/invoices:
    post:
      tags: [Payments]
      summary: Create payment invoice
      description: |
        Creates a Lightning Network BOLT11 invoice for payment. Supports multiple currencies
        with automatic conversion to satoshis.
      operationId: createInvoice
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceRequest'
            examples:
              sats:
                summary: Invoice in satoshis
                value:
                  amount: 10000
                  currency: "SAT"
                  description: "Payment for premium article access"
                  metadata:
                    contentId: "content-123"
                  expiresIn: 3600
              usd:
                summary: Invoice in USD (auto-converted)
                value:
                  amount: 5.00
                  currency: "USD"
                  description: "Monthly subscription"
                  payerInfo:
                    userId: "user-456"
                    email: "alice@example.com"
                  metadata:
                    subscriptionId: "sub-789"
                  expiresIn: 1800
      responses:
        '201':
          description: Invoice created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /payments/subscriptions:
    post:
      tags: [Payments]
      summary: Create subscription
      description: |
        Creates a new recurring subscription for a creator's content tier.
      operationId: createSubscription
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
            examples:
              monthly:
                value:
                  creatorId: "creator-123"
                  subscriberId: "user-456"
                  plan:
                    planId: "plan-789"
                    tier: "premium"
                    price: 50000
                    currency: "SAT"
                    interval: "monthly"
                  paymentMethod:
                    type: "lightning"
      responses:
        '201':
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /payments/subscriptions/{id}:
    put:
      tags: [Payments]
      summary: Update subscription
      description: |
        Updates an existing subscription (change plan, pause, or reactivate).
      operationId: updateSubscription
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Subscription ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubscriptionRequest'
      responses:
        '200':
          description: Subscription updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'

    delete:
      tags: [Payments]
      summary: Cancel subscription
      description: |
        Cancels an active subscription. Can be set to cancel immediately or at period end.
      operationId: cancelSubscription
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Subscription ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelSubscriptionRequest'
            examples:
              immediate:
                summary: Cancel immediately
                value:
                  subscriptionId: "sub-123"
                  reason: "No longer need access"
                  cancelAt: "immediately"
              period_end:
                summary: Cancel at period end
                value:
                  subscriptionId: "sub-123"
                  reason: "Switching to different tier"
                  cancelAt: "period_end"
      responses:
        '200':
          description: Subscription cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelSubscriptionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from the authentication endpoint.

        Token must include:
        - User ID
        - NOSTR public key
        - User role (supporter, creator, admin)
        - Expiration timestamp

        Example:
        ```
        Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        ```
  headers:
    X-RateLimit-Limit:
      description: Total number of requests allowed in the current window
      schema:
        type: integer
      example: 100
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current window
      schema:
        type: integer
      example: 87
    X-RateLimit-Reset:
      description: Unix timestamp when the rate limit resets
      schema:
        type: integer
      example: 1698415200
  schemas:
...
    CreateInvoiceRequest:
      type: object
      required:
        - amount
        - currency
        - description
      properties:
        amount:
          type: number
          format: float
          minimum: 0
          description: Invoice amount
          example: 10000
        currency:
          type: string
          enum: [SAT, BTC, USD, EUR, GBP]
          description: Currency code (auto-converts to SAT)
          example: "SAT"
        description:
          type: string
          minLength: 1
          maxLength: 500
          description: Invoice description
          example: "Payment for premium article access"
        payerInfo:
          type: object
          properties:
            userId:
              type: string
              format: uuid
            email:
              type: string
              format: email
            name:
              type: string
        metadata:
          type: object
          additionalProperties: true
          description: Custom metadata for the invoice
        expiresIn:
          type: integer
          minimum: 60
          maximum: 86400
          default: 3600
          description: Invoice expiration time in seconds
        webhookUrl:
          type: string
          format: uri
          description: Webhook URL for payment notifications
    InvoiceResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            invoiceId:
              type: string
              format: uuid
            paymentRequest:
              type: string
              description: BOLT11 Lightning invoice
              example: "lnbc100n1..."
            amount:
              type: object
              properties:
                requested:
                  type: number
                  format: float
                currency:
                  type: string
                sats:
                  type: integer
                  description: Amount in satoshis
            status:
              type: string
              enum: [pending, paid, expired, cancelled]
            description:
              type: string
            createdAt:
              type: string
              format: date-time
            expiresAt:
              type: string
              format: date-time
            paymentHash:
              type: string
              description: Lightning payment hash
            metadata:
              type: object
              additionalProperties: true
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
...
    CreateSubscriptionRequest:
      type: object
      required:
        - creatorId
        - subscriberId
        - plan
        - paymentMethod
      properties:
        creatorId:
          type: string
          format: uuid
        subscriberId:
          type: string
          format: uuid
        plan:
          type: object
          required:
            - planId
            - tier
            - price
            - currency
            - interval
          properties:
            planId:
              type: string
              format: uuid
            tier:
              type: string
              example: "premium"
            price:
              type: integer
              minimum: 0
            currency:
              type: string
            interval:
              type: string
              enum: [monthly, yearly]
        paymentMethod:
          type: object
          required:
            - type
          properties:
            type:
              type: string
              enum: [lightning, card]
            details:
              type: object
              additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
    SubscriptionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            subscriptionId:
              type: string
              format: uuid
            creatorId:
              type: string
              format: uuid
            subscriberId:
              type: string
              format: uuid
            plan:
              type: object
              properties:
                planId:
                  type: string
                  format: uuid
                tier:
                  type: string
                price:
                  type: integer
                currency:
                  type: string
                interval:
                  type: string
            status:
              type: string
              enum: [active, cancelled, expired, pending]
            currentPeriod:
              type: object
              properties:
                start:
                  type: string
                  format: date-time
                end:
                  type: string
                  format: date-time
            nextBillingDate:
              type: string
              format: date-time
            createdAt:
              type: string
              format: date-time
            cancelledAt:
              type: string
              format: date-time
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
    UpdateSubscriptionRequest:
      type: object
      required:
        - subscriptionId
        - updates
      properties:
        subscriptionId:
          type: string
          format: uuid
        updates:
          type: object
          properties:
            planId:
              type: string
              format: uuid
            status:
              type: string
              enum: [active, paused, cancelled]
            paymentMethod:
              type: object
              properties:
                type:
                  type: string
                details:
                  type: object
                  additionalProperties: true
    CancelSubscriptionRequest:
      type: object
      required:
        - subscriptionId
      properties:
        subscriptionId:
          type: string
          format: uuid
        reason:
          type: string
          maxLength: 1000
        cancelAt:
          type: string
          enum: [immediately, period_end]
          default: period_end
    CancelSubscriptionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            subscriptionId:
              type: string
              format: uuid
            status:
              type: string
            cancelledAt:
              type: string
              format: date-time
            effectiveDate:
              type: string
              format: date-time
              description: When cancellation takes effect
            refundAmount:
              type: integer
              description: Refund amount in satoshis (if applicable)
            refundStatus:
              type: string
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
...
    ResponseMetadata:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
          description: Unique request identifier for support/debugging
        timestamp:
          type: string
          format: date-time
          description: Server response timestamp
        processingTime:
          type: number
          format: float
          description: Request processing time in milliseconds
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          default: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code for programmatic handling
              example: "E001"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid request parameters"
            details:
              type: object
              additionalProperties: true
              description: Additional error context
            field:
              type: string
              description: Field name for validation errors
            validationErrors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  value:
                    description: Invalid value that caused error
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
  responses:
    ValidationError:
      description: Validation error - Request contains invalid data
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              value:
                success: false
                error:
                  code: "E001"
                  message: "Validation failed"
                  validationErrors:
                    - field: "title"
                      message: "Title is required"
                    - field: "amount"
                      message: "Amount must be greater than 0"
                metadata:
                  requestId: "req-123"
                  timestamp: "2025-10-27T10:30:00Z"
                  processingTime: 5.2
    Unauthorized:
      description: Unauthorized - Authentication required or token invalid
      headers:
        WWW-Authenticate:
          description: Authentication scheme
          schema:
            type: string
          example: Bearer realm="Sovren API"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing_token:
              value:
                success: false
                error:
                  code: "E401"
                  message: "Authentication required"
                  details:
                    reason: "No bearer token provided"
                metadata:
                  requestId: "req-456"
                  timestamp: "2025-10-27T10:30:00Z"
            invalid_token:
              value:
                success: false
                error:
                  code: "E402"
                  message: "Invalid or expired token"
                  details:
                    reason: "Token signature verification failed"
...
    NotFound:
      description: Not Found - Resource doesn't exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            not_found:
              value:
                success: false
                error:
                  code: "E404"
                  message: "Resource not found"
                  details:
                    resourceType: "content"
                    resourceId: "550e8400-e29b-41d4-a716-446655440000"
    RateLimitExceeded:
      description: Too Many Requests - Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
        Retry-After:
          description: Seconds until rate limit resets
          schema:
            type: integer
          example: 3600
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            rate_limit:
              value:
                success: false
                error:
                  code: "E429"
                  message: "Rate limit exceeded"
                  details:
                    limit: 100
                    window: "1 hour"
                    retryAfter: 3600
    InternalError:
      description: Internal Server Error - Something went wrong on the server
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            server_error:
              value:
                success: false
                error:
                  code: "E500"
                  message: "Internal server error"
                  details:
                    message: "An unexpected error occurred. Our team has been notified."
                metadata:
                  requestId: "req-789"
                  timestamp: "2025-10-27T10:30:00Z"